requirements.txt
cffi==1.15.1
cryptography==38.0.1
idna==3.4
pycparser==2.21
pyopenssl==22.0.0
